added verify code API

+++ database/migrations/2023_09_28_035236_create_users_verify_mobile.php
@@ -17,7 +17,7 @@ class CreateUsersVerifyMobile extends Migration
             $table->id();
             $table->unsignedBigInteger('user_id');
             $table->string('country_code');
-            $table->integer('mobile_no');
+            $table->bigInteger('mobile_no');
             $table->integer('otp');
             $table->timestamps();
